API Reference

AbstractModel.jl API reference.

model.jl

Model

AbstractModel.Initialize — Function

Initialize(input_data::OrderedDict)

Function to initialize the model based on given input data. This function calls functions initializing the solver, model, etc.

Arguments

input_data::OrderedDict: Dictionary containing data used to build the model.

Examples

'''julia-repl julia> modelcontents = Initialize(inputdata); "constraint" => ... "expression" => ... "variable" => ... "tuple" => ... "genconstraint" => ... "genexpression" => ... "res_dir" => ... "model" => ... '''

AbstractModel.solve_model — Function

solve_model(model_contents::OrderedDict)

Function to solve model.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

Examples

'''julia repl julia> Predicer.solvemodel(modelcontents); '''

Missing docstring.

Missing docstring for export_model_contents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_result_dataframe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for write_bid_matrix. Check Documenter's build log for details.

constraints.jl

tuples.jl

variables.jl

structures.jl