API Reference

AbstractModel.jl API reference.

model.jl

Model

AbstractModel.Initialize — Function

Initialize(input_data::OrderedDict)

Function to initialize the model based on given input data. This function calls functions initializing the solver, model, etc.

Arguments

input_data::OrderedDict: Dictionary containing data used to build the model.

Examples

julia> model_contents = Initialize(input_data);
OrderedDict{Any, Any} with 8 entries:
  "constraint"     => ...
  "expression"     => ...
  "variable"       => ...
  "tuple"          => ...
  "gen_constraint" => ...
  "gen_expression" => ...
  "res_dir"        => ...
  "model"          => ...

AbstractModel.solve_model — Function

solve_model(model_contents::OrderedDict)

Function to use the optimizer to solve model.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.export_model_contents — Function

export_model_contents(model_contents::OrderedDict, results::Bool)

Saves the contents of the model dict to an excel file.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.get_result_dataframe — Function

get_result_dataframe(model_contents,type="",process="",node="",scenario="")

Returns a dataframe containing specific information from the model?

Arguments

model_contents::OrderedDict: ?

type: ?

process: ?

node: ?

scenario: ?

AbstractModel.write_bid_matrix — Function

write_bid_matrix(model_contents::OrderedDict, input_data::OrderedDict)

Returns the bid matric generated by the model?

constraints.jl

Constraints

AbstractModel.create_constraints — Function

create_constraints(model_contents::OrderedDict, input_data::OrderedDict)

Create all constraints used in the model.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_node_balance — Function

setup_node_balance(model_contents::OrderedDict, input_data::OrderedDict)

Setup node balance constraints used in the model.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_process_online_balance — Function

setup_process_online_balance(model_contents::OrderedDict, input_data::OrderedDict)

Setup necessary functionalities for processes with binary online variables.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_process_balance — Function

setup_process_balance(model_contents::OrderedDict, input_data::OrderedDict)

Setup constraints used in process balance calculations.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_processes_limits — Function

setup_processes_limits(model_contents::OrderedDict, input_data::OrderedDict)

Setup constraints used for process limitations, such as min/max loads, unit starts and participation in reserves.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_reserve_balances — Function

setup_reserve_balances(model_contents::OrderedDict, input_data::OrderedDict)

Setup constraints for reserves.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_ramp_constraints — Function

setup_ramp_constraints(model_contents::OrderedDict, input_data::OrderedDict)

Setup process ramp constraints, based on ramp limits defined in input data and participation in reserves.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_fixed_values — Function

setup_fixed_values(model_contents::OrderedDict, input_data::OrderedDict)

Setup constraints for setting fixed process values at certain timesteps.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_bidding_constraints — Function

setup_bidding_constraints(model_contents::OrderedDict, input_data::OrderedDict)

Setup constraints for market bidding.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_generic_constraints — Function

setup_generic_constraints(model_contents::OrderedDict, input_data::OrderedDict)

Setup generic constraints.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_cost_calculations — Function

setup_cost_calculations(model_contents::OrderedDict, input_data::OrderedDict)

Setup expressions used for calculating the costs in the model.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.setup_objective_function — Function

setup_objective_function(model_contents::OrderedDict, input_data::OrderedDict)

Sets up the objective function, which in this model aims to minimize the costs.

tuples.jl

Tuples

Indexing of constraints, expressions and variables in this model is done using tuples. The tuples contain strings, such as ("dh", "s1", "t1").

n = node

p = process

s = scenario

t = timestep

so = process source

si = process sink

res = reserve market

rd = reserve direction (up, down, or up/down)

rt = reserve type

AbstractModel.create_tuples — Function

create_tuples(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Create all tuples used in the model, and save them in the model_contents dict.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

input_data::OrderedDict: Dictionary containing data used to build the model.

AbstractModel.create_res_nodes_tuple — Function

create_res_nodes_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates the tuple containing all the nodes which have a reserve. Form: (node).

AbstractModel.create_res_tuple — Function

create_res_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates the reserve tuple. Form: (res, n, rd, s, t).

AbstractModel.create_process_tuple — Function

create_process_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing process topology for each timestep. Form: (p, so, si, s, t).

AbstractModel.create_res_potential_tuple — Function

create_res_potential_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on reserve participation in each timestep. Form: (rd, rt, p, so, si, s, t).

AbstractModel.create_proc_online_tuple — Function

create_proc_online_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing processes with online variables for each timestep. Form: (p, s, t).

AbstractModel.create_res_pot_prod_tuple — Function

create_res_pot_prod_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on potential reserve participation per unit for each timestep. Form: (rd, rt, p, so, si, s, t).

AbstractModel.create_res_pot_cons_tuple — Function

create_res_pot_cons_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on potential reserve participation per unit for each timestep. Form: (rd, rt, p, so, si, s, t).

AbstractModel.create_node_state_tuple — Function

create_node_state_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing each node with a state (reserve) for each timestep. Form: (n, s, t).

AbstractModel.create_node_balance_tuple — Function

create_node_balance_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing nodes over which balance should be calculated. Form: (n s, t).

AbstractModel.create_proc_potential_tuple — Function

create_proc_potential_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on potential reserve participation per unit for each timestep. Form: (rd, rt, p, so, si, s, t).

AbstractModel.create_proc_balance_tuple — Function

create_proc_balance_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing all processes, over which balance is to be calculated, for each timestep. Form: (p, s, t).

AbstractModel.create_proc_op_balance_tuple — Function

create_proc_op_balance_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing all processes with piecewise efficiency, for each timestep and each operating point. Form: (p, s, t, operating_point).

AbstractModel.create_proc_op_tuple — Function

create_proc_op_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing all processes with piecewise efficiency, for each timestep. Form: (p, s, t).

AbstractModel.create_cf_balance_tuple — Function

create_cf_balance_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on processes with an capacity factor, for each timestep. Form: (p, so, si, s, t).

AbstractModel.create_lim_tuple — Function

create_lim_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple ?. Form: (p, so, si, s, t).

AbstractModel.create_trans_tuple — Function

create_trans_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing information on transport processes, for each timestep. Form: (p, so, si, s, t).

AbstractModel.create_res_eq_tuple — Function

create_res_eq_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple with each node with reserves, for relevant reserve type and each timestep. Form: (n, rt, s, t).

AbstractModel.create_res_eq_updn_tuple — Function

create_res_eq_updn_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing all (reserve) markets with an (up/down) direction. Form: (m, s, t).

AbstractModel.create_res_final_tuple — Function

create_res_final_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing all (reserve) markets for each timestep. Form: (m, s, t).

AbstractModel.create_fixed_value_tuple — Function

create_fixed_value_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing timesteps containing fixed market states. Form: (m, s, t).

AbstractModel.create_ramp_tuple — Function

create_ramp_tuple(model_contents::OrderedDict, input_data::OrderedDict::OrderedDict)

Creates tuple containing timesteps with ramp possibility. Form: (p, so, si, s, t).

variables.jl

Variables

AbstractModel.create_variables — Function

create_variables(model_contents::OrderedDict)

Create the variables used in the model, and save them in the model_contents dict.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.create_v_flow — Function

create_v_flow(model_contents::OrderedDict)

Sets up v_flow, which is the variable symbolising flows between nodes/processes.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.create_v_online — Function

create_v_online(model_contents::OrderedDict)

Sets up variables used for modelling functionality for processes with binary online variables, as well as starts/stops.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.create_v_reserve — Function

create_v_reserve(model_contents::OrderedDict)

Sets up the variables used for modelling reserves.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.create_v_state — Function

create_v_state(model_contents::OrderedDict)

Sets up variables used for node state (storage) functionality.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

AbstractModel.create_v_flow_op — Function

create_v_flow_op(model_contents::OrderedDict)

Sets up variables for processes with piecewise efficiency functionality.

Arguments

model_contents::OrderedDict: Dictionary containing all data and structures used in the model.

structures.jl

Structures

AbstractModel.Node — Type

struct Node
    name::String
    is_commodity::Bool
    is_state::Bool
    is_res::Bool
    is_inflow::Bool
    is_market::Bool
    state::State
    cost::Vector{TimeSeries}
    inflow::Vector{TimeSeries}
    function Node(name, is_commodity, is_state, is_res, is_inflow, is_market, state_max, in_max, out_max, initial_state)
        return new(name, is_commodity, is_state, is_res, is_inflow, is_market, State(in_max, out_max, initial_state, state_max), [], [])
    end
end

A struct for nodes.

Fields

name::String: Name of the node.

is_commodity::Bool: Flag indicating of the node is a commodity.

is_state::Bool: Flag indicating of the node has a state (storage).

is_res::Bool: Flag indicating of the node participates as a reserve.

is_inflow::Bool: Flag indicating of the node has a inflow.

is_market::Bool: Flag indicating of the node is a market node.

state::State: The state of the node.

cost::Vector{TimeSeries}: Vector containing TimeSeries with the costs for each scenario.

inflow::Vector{TimeSeries}: Vector contining TimeSeries with the inflows for each scenario.

AbstractModel.Process — Type

struct Process
    name::String
    is_cf::Bool
    is_cf_fix::Bool
    is_online::Bool
    is_res::Bool
    eff::Float64
    conversion::Integer
    load_min::Float64
    load_max::Float64
    start_cost::Float64
    min_online::Int64
    min_offline::Int64
    initial_state::Bool
    topos::Vector{Topology}
    cf::Vector{TimeSeries}
    eff_ts::Vector{TimeSeries}
    eff_ops::Vector{Any}
    eff_fun::Vector{Tuple{Any,Any}}
    function Process(name, is_cf, is_cf_fix, is_online, is_res, eff, conversion, load_min, load_max, start_cost, min_online, min_offline, initial_state)
        return new(name, is_cf, is_cf_fix, is_online, is_res, eff, conversion, load_min, load_max, start_cost, min_online, min_offline, initial_state, [], [], [], [], [])
    end
end

A struct for a process (unit).

Fields

name::String: Name of the process.

is_cf::Bool: Flag indicating if the process is a cf (capacity factor) process, aka depends on a TimeSeries.

is_cf_fix::Bool: Flag indicating if the cf TimeSeries is a upper limit (false) or a set value (true).

is_online::Bool: Flag indicating if the process has an binary online variable.

is_res::Bool: Flag indicating if the process can participate in a reserve.

eff::Float64: Process conversion efficiency.

conversion::Integer:

load_min::Float64: Minimum allowed load over the process, min 0, max 1.

load_max::Float64: Maximum allowed load over the process, min 0, max 1.

start_cost::Float64: Cost to start the process, if the 'is_online' flag is true.

min_online::Int64: Minimum time the process has to be online after start.

min_offline::Int64: Minimum time the process has to be offline after start.

initial_state::Bool: Initial state (on/off) of the process at the start of simulation.

topos::Vector{Topology}: Vector containing the topologies of the process.

cf::Vector{TimeSeries}: Vector containing TimeSeries limiting a cf process.

eff_ts::Vector{TimeSeries}: Vector of TimeSeries containing information on efficiency depending on time.

eff_ops::Vector{Any}: Vector containing operating points for a piecewise efficiency function.

eff_fun::Vector{Tuple{Any,Any}}: Vector containing efficiencies for a piecewise efficiency function.

AbstractModel.TimeSeries — Type

mutable struct TimeSeries
    scenario::Any
    series::Vector{Tuple{Any, Any}}
    function TimeSeries(scenario="", series=0)
        if series != 0
            return new(scenario, series)
        else
            return new(scenario, [])
        end
    end
end

A struct for time series. Includes linked scenario and a vector containing tuples of time and value.

AbstractModel.State — Type

struct State
    in_max::Float64
    out_max::Float64
    state_max::Float64
    state_min::Float64
    initial_state::Float64
    state_loss::Float64
    function State(in_max, out_max, initial_state, state_max, state_loss, state_min=0)
        return new(in_max, out_max, state_max, state_min, initial_state, state_loss)
    end
end

A struct for node states (storage), holds information on the parameters of the state.

Fields

in_max::Float64: Value for maximum increase of state variable value between timesteps.

out_max::Float64: Value for maximum decrease of state variable value between timesteps.

state_max::Float64: Maximum value for state variable.

state_min::Float64: Minimum value for state variable.

initial_state::Float64: Initial value of the state variable at t = 0.

state_loss: Losses over time in the state.

AbstractModel.Market — Type

struct Market
    name::String
    type::String
    node::Any
    direction::String
    realisation::Float64
    reserve_type::String
    price::Vector{TimeSeries}
    fixed::Vector{Tuple{Any,Any}}
    function Market(name, type, node, direction, realisation, reserve_type)
        return new(name, type, node, direction, realisation, reserve_type, [], [])
    end
end

A struct for markets.

Fields

name::String: Name of the market.

type::String: Type of the market (energy/reserve).

node::Any: Name of the node this market is connected to.

direction::String: Direction of the market (up/down/updown).

realisation::Float64: Realisation probability.

reserve_type::String: Type of the reserve market.

price::Vector{TimeSeries}: Vector containing TimeSeries of the market price in different scenarios.

fixed::Vector{Tuple{Any,Any}}: Vector containing information on the market being fixed.

AbstractModel.Topology — Type

struct Topology
    source::String
    sink::String
    capacity::Float64
    VOM_cost::Float64
    ramp_up::Float64
    ramp_down::Float64
    function Topology(source, sink, capacity, VOM_cost, ramp_up, ramp_down)
        return new(source, sink, capacity, VOM_cost, ramp_up, ramp_down)
    end
end

A struct for a process topology, signifying the connection between flows in a process.

Fields

source::String: Name of the source of the topology.

sink::String: Name of the sink of the topology.

capacity::Float64: Upper limit of the flow variable for the topology.

VOM_cost::Float64: VOM cost of using this connection.

ramp_up::Float64: Maximum allowed increase of the linked flow variable value between timesteps.

ramp_down::Float64: Minimum allowed increase of the linked flow variable value between timesteps.

AbstractModel.ConFactor — Type

struct ConFactor
    flow::Tuple{Any,Any}
    data::Vector{TimeSeries}
    function ConFactor(flow,data)
        return new(flow,data)
    end
end

Struct for general constraints factors.

Fields

flow::Tuple{Any,Any}: ??

data::Vector{TimeSeries}: ??

AbstractModel.GenConstraint — Type

struct GenConstraint
    name::String
    type::String
    factors::Vector{ConFactor}
    constant::Vector{TimeSeries}
    function GenConstraint(name,type)
        return new(name,type,[],[])
    end
end

Struct for general constraints.

Fields

name::String: Name of the generic constraint.

type::String: Type of the generic constraint.

factors::Vector{ConFactor}: Vector of ConFactors.

constant::Vector{TimeSeries}: TimeSeries?